// src/app/shared/model/product.model.ts
import { Child } from './child.model';
import { ProductImage } from './productImage.model';
export class Product {
    _id: string;
    productName: string;
    manufactureInfo: string;
    vendorId: string;
    variantId: string;
    sizeVariantId: string;
    costIncludes: string;
    variantTypeId: string;
    colorId: string;
    brandName: string;
    hsnCode: string;
    gstIn: string;
    sku: string;
    styleCode: string;
    variation: string;
    productsType: string;
    packOf: number;
    size: string;
    weight: string;
    fabric: string;
    washCare: string;
    productDimension: string;
    packingDimension: string;
    discount: number;
    sp: number;
    vp: number;
    mrp: number;
    closure: string;
    pattern: string;
    productDescription: string;
    toFitWaist: string;
    searchTerms1: string;
    searchTerms2: string;
    searchTerms3: string;
    searchTerms4: string;
    searchTerms5: string;
    superCategoryId: string;
    superCategoryName: string;
    catalogueName: string;
    mainCategoryId: string;
    mainCategoryName: string;
    subCategoryId: string;
    subCategoryName: string;
    color: string;z
    tags: string;
    metaTitle: string;
    metaDescription: string;
    metaKeyword: string;
    dateAdded: Date;
    dateModified: Date;
    brandId: string;
    variantType: string;
    publish: boolean;
    sizeName: string;
    sizeId: string;
    productImage: [ProductImage];
    ttsPortol: number;
    ttsVendor: number;
    child: [Child];
    nameSort: string;
    dateSort: string;
    sizeVariant: string;
// extra
    bulletPoints: string;
    height: string;
    material: string;
    occassion: string;
    manufacturer: string;
    price: string;
    headChild: boolean;
    quantity: string;
    length: string;
    breath: string;
    hsn: string;
    sizeGuideId: string;
    guideEdit: boolean;
    sizeGuideName: string;
    result: string;
}